requirements.txt
asgiref==3.8.1
Django==5.1.1
django-session-timeout==0.1.0
psycopg==3.2.1
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
